+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/StatisticsEndpoint.java
@@ -27,10 +27,10 @@ import static org.opencastproject.util.doc.rest.RestParameter.Type.STRING;
 
 import org.opencastproject.adminui.impl.ProviderQuery;
 import org.opencastproject.adminui.impl.RawProviderQuery;
-import org.opencastproject.adminui.index.AdminUISearchIndex;
 import org.opencastproject.elasticsearch.api.SearchIndexException;
-import org.opencastproject.elasticsearch.index.event.Event;
-import org.opencastproject.elasticsearch.index.series.Series;
+import org.opencastproject.elasticsearch.index.ElasticsearchIndex;
+import org.opencastproject.elasticsearch.index.objects.event.Event;
+import org.opencastproject.elasticsearch.index.objects.series.Series;
 import org.opencastproject.index.service.api.IndexService;
 import org.opencastproject.security.api.Organization;
 import org.opencastproject.security.api.SecurityService;
@@ -102,7 +102,7 @@ public class StatisticsEndpoint {
 
   private SecurityService securityService;
   private IndexService indexService;
-  private AdminUISearchIndex searchIndex;
+  private ElasticsearchIndex searchIndex;
   private StatisticsService statisticsService;
   private StatisticsExportService statisticsExportService;
 
@@ -117,7 +117,7 @@ public class StatisticsEndpoint {
   }
 
   @Reference
-  public void setSearchIndex(AdminUISearchIndex searchIndex) {
+  public void setSearchIndex(ElasticsearchIndex searchIndex) {
     this.searchIndex = searchIndex;
   }