+++ modules/statistics-export-service-impl/src/main/java/org/opencastproject/statistics/export/impl/StatisticsExportServiceImpl.java
@@ -28,11 +28,11 @@ import org.opencastproject.elasticsearch.api.SearchIndexException;
 import org.opencastproject.elasticsearch.api.SearchQuery;
 import org.opencastproject.elasticsearch.api.SearchResult;
 import org.opencastproject.elasticsearch.api.SearchResultItem;
-import org.opencastproject.elasticsearch.index.AbstractSearchIndex;
-import org.opencastproject.elasticsearch.index.event.Event;
-import org.opencastproject.elasticsearch.index.event.EventSearchQuery;
-import org.opencastproject.elasticsearch.index.series.Series;
-import org.opencastproject.elasticsearch.index.series.SeriesSearchQuery;
+import org.opencastproject.elasticsearch.index.ElasticsearchIndex;
+import org.opencastproject.elasticsearch.index.objects.event.Event;
+import org.opencastproject.elasticsearch.index.objects.event.EventSearchQuery;
+import org.opencastproject.elasticsearch.index.objects.series.Series;
+import org.opencastproject.elasticsearch.index.objects.series.SeriesSearchQuery;
 import org.opencastproject.index.service.api.IndexService;
 import org.opencastproject.mediapackage.MediaPackage;
 import org.opencastproject.metadata.dublincore.DublinCoreMetadataCollection;
@@ -183,7 +183,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
       Instant from,
       Instant to,
       DataResolution dataResolution,
-      AbstractSearchIndex index,
+      ElasticsearchIndex index,
       ZoneId zoneId
   ) throws SearchIndexException, UnauthorizedException, NotFoundException {
     if (!(provider instanceof TimeSeriesProvider)) {
@@ -229,7 +229,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
 
   @Override
   public String getCSV(StatisticsProvider provider, String resourceId, Instant from, Instant to, DataResolution
-          dataResolution, AbstractSearchIndex index, ZoneId zoneId, boolean fullMetadata, DetailLevel detailLevel,
+          dataResolution, ElasticsearchIndex index, ZoneId zoneId, boolean fullMetadata, DetailLevel detailLevel,
           int limit, int offset, Map<String, String> filters)
           throws SearchIndexException, UnauthorizedException, NotFoundException {
     if (!(provider instanceof TimeSeriesProvider)) {
@@ -286,7 +286,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
       Instant from,
       Instant to,
       DataResolution dataResolution,
-      AbstractSearchIndex index,
+      ElasticsearchIndex index,
       ZoneId zoneId,
       CSVPrinter printer,
       boolean fullMetaData,
@@ -310,7 +310,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
   }
 
   private void printSeries(StatisticsProvider provider, String resourceId, Instant from, Instant to,
-                           DataResolution dataResolution, AbstractSearchIndex index, ZoneId zoneId, CSVPrinter printer,
+                           DataResolution dataResolution, ElasticsearchIndex index, ZoneId zoneId, CSVPrinter printer,
                            boolean fullMetadata, int limit, int offset)
           throws SearchIndexException, NotFoundException, IOException {
     if (offset != 0) {
@@ -335,7 +335,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
       Instant from,
       Instant to,
       DataResolution dataResolution,
-      AbstractSearchIndex index,
+      ElasticsearchIndex index,
       ZoneId zoneId,
       CSVPrinter printer,
       boolean fullMetadata,
@@ -398,7 +398,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
       Instant from,
       Instant to,
       DataResolution dataResolution,
-      AbstractSearchIndex index,
+      ElasticsearchIndex index,
       ZoneId zoneId,
       CSVPrinter printer,
       boolean fullMetadata,
@@ -444,7 +444,7 @@ public class StatisticsExportServiceImpl implements StatisticsExportService, Man
       Instant from,
       Instant to,
       DataResolution dataResolution,
-      AbstractSearchIndex index,
+      ElasticsearchIndex index,
       ZoneId zoneId,
       CSVPrinter printer,
       boolean fullMetadata,