+++ modules/workflow-service-impl/src/test/java/org/opencastproject/workflow/impl/WorkflowServiceImplAuthzTest.java
@@ -28,8 +28,8 @@ import static org.opencastproject.workflow.impl.SecurityServiceStub.DEFAULT_ORG_
 
 import org.opencastproject.assetmanager.api.AssetManager;
 import org.opencastproject.elasticsearch.api.SearchResult;
-import org.opencastproject.elasticsearch.index.AbstractSearchIndex;
-import org.opencastproject.elasticsearch.index.event.EventSearchQuery;
+import org.opencastproject.elasticsearch.index.ElasticsearchIndex;
+import org.opencastproject.elasticsearch.index.objects.event.EventSearchQuery;
 import org.opencastproject.mediapackage.MediaPackage;
 import org.opencastproject.mediapackage.MediaPackageBuilderFactory;
 import org.opencastproject.metadata.api.MediaPackageMetadataService;
@@ -246,13 +246,12 @@ public class WorkflowServiceImplAuthzTest {
 
     SearchResult result = EasyMock.createNiceMock(SearchResult.class);
 
-    final AbstractSearchIndex index = EasyMock.createNiceMock(AbstractSearchIndex.class);
+    final ElasticsearchIndex index = EasyMock.createNiceMock(ElasticsearchIndex.class);
     EasyMock.expect(index.getIndexName()).andReturn("index").anyTimes();
     EasyMock.expect(index.getByQuery(EasyMock.anyObject(EventSearchQuery.class))).andReturn(result).anyTimes();
     EasyMock.replay(result, index);
 
-    service.setAdminUiIndex(index);
-    service.setExternalApiIndex(index);
+    service.setIndex(index);
 
     // Activate
     service.activate(null);